Add snap-together pegs to modularise the design
The pegs are generated using a generalised function which either generates the solid peg or a hole, depending on a boolean argument to the module. The idea is that for simplicity, creating the locking holes for a set of pegs should use exactly the same parameters to the snap_pegs module as the solid pegs, with only the boolean argument being different.