Tweaks to EOS switch+others after switch burnout
These are an attempt to stop the flipper from machine-gunning instead of holding. The theory is that because the flipper clips the other half of the switch as it flips due to miscalculated distances, doesn't function correctly. The EOS switch pegs moved by a couple of mm which should in theory fix it.